Fix highlighting bug in macro arguments
Previously, the TO-ARGS-* contexts ended if the next character was a
comma or closing parenthesis; but they need to extend from the beginning
comma to the closing parenthesis so that they contain the next TO-ARGS-*
subcontext instead of matching themselves more than once.

e.g. if AC_SOMETHING was a "MMSS" macro

AC_SOMETHING([foo], [bar], [false], [true])
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ ARGS-MMSS
             ^^^^^ ARG-M
                  ^^^^^^^ TO-ARGS-MSS
                    ^^^^^ ARG-M
                         ^^^^^^^^^ TO-ARGS-MSS
                           ^^^^^^^ ARG-M
                                  ^^^^^^^^ TO-ARGS-MSS
                                    ^^^^^^ ARG-M

should be

AC_SOMETHING([foo], [bar], [false], [true])
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ ARGS-MMSS
             ^^^^^ ARG-M
                  ^^^^^^^^^^^^^^^^^^^^^^^^ TO-ARGS-MSS
                    ^^^^^ ARG-M
                         ^^^^^^^^^^^^^^^^^ TO-ARGS-SS
                           ^^^^^^^ ARG-S
                                  ^^^^^^^^ TO-ARGS-S
                                    ^^^^^^ ARG-S

This fixed it so that extra (illegal) arguments are now highlighted as
such. Also, previously, shell (S) or pure M4 (M) arguments could be
highlighted wrongly as shown in the diagram above.