Remove Makefile2 from README
When the Makefile2 syntax was written, it was better than the built-in
Makefile highlighting in Sublime Text. That is no longer the case. So,
it's unlikely that anyone will want to override the built-in Makefile
syntax with Makefile2. Remove these instructions from the README.

Instead, people will probably want to override the .am extension with
the Automake syntax, because the newer Makefile highlighting in Sublime
Text opens Makefile.am as Makefile syntax by default. Add instructions
for this.